Final version with delete files

+++ OS.sh
@@ -14,29 +14,38 @@ if [[ "$1" = "--help" ]]
 then
     echo "Use ./OS.sh [adres from] [type] [adres to] -c [count of backups of this type]"
     echo
-    echo "Не паникуй, ща всё будет"
-    echo "Смотри, вбиваешь полный адрес места, откуда хочешь копировать,"
-    echo "потом буквы расширения и после этого полный адрес места,"
-    echo "куда хочешь капировать. Не забудь после адреса поставить /"
-    echo "Пример: ./OS.sh Desktop/OSi/in/ docx Desktop/ -c 5"
+    echo "Эта программа позволяет создавать резервные копии файлов с заданным расширением."
+    echo "Для создания копии необходипо обязательно указать папку из которой нужно сохранить файлы, нужное расширение файлов, и папку, в которую сохранить резервную копию."
+    echo "Опционально можно указать количество активных резервных копий, при этом старые копии удалятся."
+    echo "Пример для создания резервных копий файлов ворд: ./OS.sh ./Desktop/OSi/in/ docx ./Desktop/"
+    echo "Пример для создания резервных копий файлов ворд с сохранением последних пяти: ./OS.sh ./Desktop/OSi/in/ docx ./Desktop/ -c 5"
+    exit 0
 else
+
+if [[ "$1" = "" ]] || [[ "$2" = "" ]] || [[ "$3" = "" ]]
+then
+    echo "Введена неверная команда"
+    echo "Используйте ./OS.sh --help для вывода информации и инструкции"
+    exit 2
+fi
 
 while [[ "$md5From" != "$md5To" ]]
 do
     date=$(date +"%H:%M:%S %d.%m.%Y")
     name="$adresTo""Backup of [""$ras""] ""$date"".bac"
+
     # create copy in zip and replace it in directory
     cd "$adresFrom" || exit
-    zip -9q temporary.zip $(find . -name "*""$ras")
+    zip -9q "temporary.zip" $(find . -name "*""$ras")
     cd || exit
     cp "$adresFrom""temporary.zip" "$name"
     rm "$adresFrom""temporary.zip"
 
     # create and chek md5sum
-    md5From=($(md5sum "$adresFrom"*".""$ras"|sort))
+    IFS=" " read -r -a md5From <<< "$(md5sum "$adresFrom"*".""$ras"|sort)"
     mkdir "/home/kali/temporary"
     unzip -q "$name" -d "/home/kali/temporary/"
-    md5To=($(md5sum "/home/kali/temporary/"*|sort))
+    IFS=" " read -r -a md5To <<< "$(md5sum "/home/kali/temporary/"*|sort)"
     rm -Rf "/home/kali/temporary/"
     ((i++)) || true
     echo "Попытка номер: ""$i"
@@ -55,8 +64,10 @@ done
 if [[ $i -le 10 ]]
 then
     echo "Всё хорошо, мы успешно сделали сохранение в файл /""$name"
+    
 else
     echo "У меня не получилось сделать сохранение"
+    exit 1
 fi
 
 
@@ -66,16 +77,17 @@ then
     if [[ "$5" = "" ]]
     then
         echo
-        echo "Вы должны вписать количество резервных копий для удаления старых"
-        echo "Use ./OS.sh --help for information"
+        echo "Вы должны указать количество резервных копий для удаления старых"
+        echo "Используйте ./OS.sh --help для вывода информации и инструкции"
+        exit 2
     else
         maxCount=$5
         count=$(ls "$adresTo""Backup of [""$ras""] "*".bac" | wc -l)
-        ((count=$count-$maxCount))
+        ((count=count-maxCount))
         i=$count
         for file in "$adresTo""Backup of [""$ras""] "*".bac"
         do
-            if [[ $count  > 0 ]]
+            if [[ $count  -gt 0 ]]
             then    
                 ((count--))
                 rm "$file"
@@ -84,8 +96,9 @@ then
             fi
         done
         echo "Успешно удалено ""$i"" устаревших копий"
+        exit 0
     fi
 fi
-
+exit 0
 
 fi
